Support for Tessellation (#607)
* Fix typo OuptutTopologyAttribute -> OutputTopologyAttribute
First pass support for handing tesselation shaders - domain and hull.

* Added attribute PatchConstantFuncAttribute
* Added  visitHLSLPatchType(HLSLPatchType* type) such that the patch type template parameters are handled
* Added IRNotePatchConstantFunc - such that the patch constant function is referenced within IR
* Added support for outputing typical tesselation attributes (although minimal validation is performed)
* Added findFunctionDeclByName

* Small improvements to diagnostic.

* Improved diagnostics and checking for geometry shader attributes.

* Added diagnostic if patchconstantfunc is not found
Handle assert failure when outputing a domain shader alone and therefore attr->patchConstantFuncDecl is not set.

* Simple script tess.hlsl to test out domain/hull shaders.

* Added url for where hull shader attributes are defined.

* Fix unsigned/signed comparison warning.

* Restore removal of fix in "Improve generic argument inference for builtins (#598)"

* Update tessellation test case to compare against fxc

The test was previously comparing against fixed expected DXBC output, but this caused problems when the test runner tried to execute the test on Linux (where there is no fxc to invoke...), and would also be a potential source of problems down the road if different users run using different builds of fxc.

The simple solution here is to convert the test to compare against fxc output generated on the fly. That test type is already filtered out on non-Windows builds, so it eliminates the portability issue (in a crude way).

I also changed the test to compile both entry points in one compiler invocation, just to streamline things into fewer distinct tests.

* Eliminate unnecessary call to `lowerFuncDecl`

In a very obscure case this could cause a bug, if the patch-constant function had somehow already been lowered (because it was called somewhere else in the code).

The call should not be needed because `ensureDecl` will lower a declaration on-demand if required, so eliminating it causes no problems for code that wouldn't be in that extreme corner case.